reorg the headers for a cleaning design
